+++ roqoqo-qryd/tests/integration/tweezer_devices.rs
@@ -487,6 +487,15 @@ fn test_change_device() {
     assert_eq!(device.current_layout, "Test");
 }
 
+/// Test TweezerDevice allow_reset field
+#[test]
+fn test_allow_reset() {
+    let mut device = TweezerDevice::new(None, None, None);
+    assert!(!device.allow_reset);
+    device.set_allow_reset(true);
+    assert!(device.allow_reset);
+}
+
 /// Test TweezerDevice change_device() method with PragmaShiftQubitsTweezers
 #[test]
 fn test_change_device_shift() {
@@ -573,6 +582,7 @@ fn test_change_device_shift() {
 fn test_from_api() {
     let mut returned_device_default = TweezerDevice::new(None, None, None);
     returned_device_default.set_tweezer_single_qubit_gate_time("PauliX", 0, 0.23, None);
+    returned_device_default.device_name = "qryd_emulator".to_string();
     let mut server = Server::new();
     let port = server
         .url()
@@ -596,22 +606,21 @@ fn test_from_api() {
 
     let response = TweezerDevice::from_api(None, None, Some(port.clone()), None, None, None);
     assert!(response.is_ok());
-
     let device = response.unwrap();
     assert_eq!(device, returned_device_default);
+    assert_eq!(device.qrydbackend(), "qryd_emulator".to_string());
 
     let response_new_seed =
         TweezerDevice::from_api(None, None, Some(port.clone()), Some(42), None, None);
     mock.assert();
     assert!(response_new_seed.is_ok());
-
     let device_new_seed = response_new_seed.unwrap();
     assert_eq!(device_new_seed.seed(), Some(42));
 
     mock.remove();
     mock = server
         .mock("GET", mockito::Matcher::Any)
-        .with_status(400)
+        .with_status(500)
         .create();
 
     let response = TweezerDevice::from_api(None, None, Some(port), None, None, None);
@@ -620,7 +629,7 @@ fn test_from_api() {
     assert_eq!(
         response.unwrap_err(),
         RoqoqoBackendError::NetworkError {
-            msg: format!("Request to server failed with HTTP status code {:?}.", 400),
+            msg: format!("Request to server failed with HTTP status code {:?}.", 500),
         }
     );
 }