+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Tracks qoqo-qryd changes after 0.5
 
+# 0.11.5
+
+* Modified `TweezerDevice.from_api()` endpoint, default device name for better `APIBackend` support
+* Added support for `PragmaActiveReset` for `APIBackend`, `TweezerDevice`
+
 # 0.11.4
 
 * Added `api_version`, `seed`, `dev` parameters to `TweezerDevice.from_api()`